requirements.txt
typer
requests
tqdm
python-decouple
datasets
pydantic
langchain
scikit-learn
